Plugin concept
This is a proposd change to how plugins are handled. In this example, alive and file_checker are loaded by simply adding the name of the plugin to a macro invocation. If the plugin identity isn't in the macro invocation, the plugin is not compiled into the binary. 

This works by returning a vector of trait objects. In order to make the these object safe, the new function is not a module level function rather than belonging to the trait.

To finish the work on this concept, arbitrate will need to be modified since we now longer have the exact plugin type. One neat thing about this setup is now that we have all of our plugins in one place, we can stop busy waiting. Instead we can figure out the shortest time until the next plugin is to be run, and wait that long. We may even be able to work with futures in the future, but I have no experience with futures and trait objects.