Initial
Added some code generation back in, but simpilier than the initial build script and doesn't touch 'main code'. 

In this concept, to enable a plugin you add it to agent/plugins/Cargo.toml as a dependency. 

The benifit to this PoC is that agent_lib can be hosted on crates.io and other plugin authors can very easily depend on it. Additionally, plugins themselves may be hosted on crates or on github. Cargo itself will handle dependency management and version control.